requirements.txt
aiogram>=3.4,<4
gspread>=6.0
google-auth>=2.0
python-dotenv>=1.0
